allow partial-dollar donations

+++ batch.py
@@ -53,7 +53,7 @@ def amount_to_charge(entry):
     Stripe wants the amount to charge in cents. So we multiply by 100 and
     return that.
     """
-    amount = int(entry['Amount'])
+    amount = float(entry['Amount'])
     if entry['Stripe_Agreed_to_pay_fees__c']:
         fees = amount * .029 + .30
     else:
